Add new fontconvert stuff with custom modifications and japanese font

+++ lib/MAKERphone/src/TFT_eSPI/Fonts/GFXFF/gfxfont.h
@@ -21,8 +21,8 @@ typedef struct { // Data stored PER GLYPH
 typedef struct { // Data stored for FONT AS A WHOLE:
 	uint8_t  *bitmap;      // Glyph bitmaps, concatenated
 	GFXglyph *glyph;       // Glyph array
-	uint8_t   first, last; // ASCII extents
-	uint8_t   yAdvance;    // Newline distance (y axis)
+	uint16_t first, last; // ASCII extents
+	uint8_t yAdvance;    // Newline distance (y axis)
 } GFXfont;
 
 #endif // LOAD_GFXFF
